fold DoG pyramid code into harris_laplace_detector.cpp
