--- modules/xfeatures2d/src/gaussian_pyramid.hpp
@@ -1,109 +1,0 @@
-/*IMPORTANT: READ BEFORE DOWNLOADING, COPYING, INSTALLING OR USING.
-
- By downloading, copying, installing or using the software you agree to this license.
- If you do not agree to this license, do not download, install,
- copy or use the software.
-
-
- License Agreement
- For Open Source Computer Vision Library
-
- Copyright (C) 2000-2008, Intel Corporation, all rights reserved.
- Copyright (C) 2008-2010, Willow Garage Inc., all rights reserved.
- Third party copyrights are property of their respective owners.
-
- Redistribution and use in source and binary forms, with or without modification,
- are permitted provided that the following conditions are met:
-
- * Redistribution's of source code must retain the above copyright notice,
- this list of conditions and the following disclaimer.
-
- * Redistribution's in binary form must reproduce the above copyright notice,
- this list of conditions and the following disclaimer in the documentation
- and/or other materials provided with the distribution.
-
- * The name of the copyright holders may not be used to endorse or promote products
- derived from this software without specific prior written permission.
-
- This software is provided by the copyright holders and contributors "as is" and
- any express or implied warranties, including, but not limited to, the implied
- warranties of merchantability and fitness for a particular purpose are disclaimed.
- In no event shall the Intel Corporation or contributors be liable for any direct,
- indirect, incidental, special, exemplary, or consequential damages
- (including, but not limited to, procurement of substitute goods or services;
- loss of use, data, or profits; or business interruption) however caused
- and on any theory of liability, whether in contract, strict liability,
- or tort (including negligence or otherwise) arising in any way out of
- the use of this software, even if advised of the possibility of such damage.*/
-
-#ifndef PYRAMID_H_
-#define PYRAMID_H_
-#include "opencv2/core.hpp"
-#include "opencv2/imgproc.hpp"
-
-namespace cv {
-
-class Pyramid
-{
-
-protected:
-    class Octave
-    {
-    public:
-        std::vector<Mat> layers;
-        Octave();
-        Octave(std::vector<Mat> layers);
-        virtual ~Octave();
-        std::vector<Mat> getLayers();
-        Mat getLayerAt(int i);
-    };
-
-    class DOGOctave
-    {
-    public:
-        std::vector<Mat> layers;
-
-        DOGOctave();
-        DOGOctave(std::vector<Mat> layers);
-        virtual ~DOGOctave();
-        std::vector<Mat> getLayers();
-        Mat getLayerAt(int i);
-    };
-
-private:
-    std::vector<Octave> octaves;
-    std::vector<DOGOctave> DOG_octaves;
-    void build(const Mat& img, bool DOG);
-public:
-    class Params
-    {
-    public:
-        int octavesN;
-        int layersN;
-        float sigma0;
-        int omin;
-        float step;
-        Params();
-        Params(int octavesN, int layersN, float sigma0, int omin);
-        void clear();
-    };
-    Params params;
-
-    Pyramid();
-    Pyramid(const Mat& img, int octavesN, int layersN = 2, float sigma0 = 1, int omin = 0,
-            bool DOG = false);
-    Mat getLayer(int octave, int layer);
-    Mat getDOGLayer(int octave, int layer);
-    float getSigma(int octave, int layer);
-    float getSigma(int layer);
-
-    virtual ~Pyramid();
-    Params getParams();
-    void clear();
-    bool empty();
-};
-
-}  // namespace cv
-
-
-#endif /* PYRAMID_H_ */
